MLDB-1433 move bucket structures out of dataset.h
