+++ core/dataset.cc
@@ -17,6 +17,7 @@
 #include "mldb/jml/utils/lightweight_hash.h"
 #include "mldb/server/dataset_context.h"
 #include "mldb/server/per_thread_accumulator.h"
+#include "mldb/server/bucket.h"
 #include "mldb/jml/utils/environment.h"
 #include "mldb/ml/jml/buckets.h"
 #include "mldb/base/parallel.h"
@@ -210,317 +211,6 @@ PersistentDatasetConfigDescription()
 
     addField("dataFileUrl", &PersistentDatasetConfig::dataFileUrl,
              "URL of the data file from which to load the dataset.");
-}
-
-/*****************************************************************************/
-/* BUCKET LIST                                                               */
-/*****************************************************************************/
-
-void
-WritableBucketList::
-init(size_t numElements, uint32_t numBuckets)
-{
-    this->numBuckets = numBuckets;
-    entryBits = ML::highest_bit(numBuckets) + 1;
-
-    // Take a number of bits per entry that evenly divides into
-    // 64 bits.
-    if (entryBits == 0) ;
-    else if (entryBits == 1) ;
-    else if (entryBits == 2) ;
-    else if (entryBits <= 4)
-        entryBits = 4;
-    else if (entryBits <= 8)
-        entryBits = 8;
-    else if (entryBits <= 16)
-        entryBits = 16;
-    else entryBits = 32;
-
-    //cerr << "using " << entryBits << " bits for " << numBuckets
-    //     << " buckets" << endl;
-
-    size_t numWords = (entryBits * numElements + 63) / 64;
-    auto writableStorage = makeSharedArray<uint64_t>(numWords);
-    this->current = writableStorage.get();
-    this->storage = writableStorage;
-    this->bitsWritten = 0;
-    this->numEntries = numElements;
-    this->numWritten = 0;
-}
-
-
-/*****************************************************************************/
-/* BUCKET DESCRIPTIONS                                                       */
-/*****************************************************************************/
-
-uint32_t
-BucketDescriptions::
-getBucket(const CellValue & val) const
-{
-    switch (val.cellType()) {
-    case CellValue::EMPTY:
-        if (!hasNulls)
-            throw HttpReturnException(500, "Unknown type for bucket");
-        return 0;
-
-    case CellValue::INTEGER:
-    case CellValue::FLOAT:
-        return numeric.getBucket(val.toDouble());
-    case CellValue::ASCII_STRING:
-    case CellValue::UTF8_STRING:
-        return strings.getBucket(val);
-    case CellValue::BLOB:
-        return blobs.getBucket(val);
-    case CellValue::TIMESTAMP:
-        return timestamps.getBucket(val);
-    case CellValue::TIMEINTERVAL:
-        return intervals.getBucket(val);
-
-    case CellValue::NUM_CELL_TYPES:
-        break;
-    }
-
-    throw HttpReturnException(500, "Unknown CellValue type for getBucket()");
-}
-
-#if 0
-CellValue
-BucketDescriptions::
-getValue(uint32_t bucket) const
-{
-    if (bucket < numeric.offset)
-        return 0;  // must be a null
-    else if (bucket < strings.offset)
-        return numeric.getValue(bucket);
-    else if (bucket < blobs.offset)
-        return blobs.getValue(bucket);
-    else if (bucket < timestamps.offset)
-        return timestamps.getValue(bucket);
-    else return intervals.getValue(bucket);
-}
-#endif
-
-uint32_t
-NumericValues::
-getBucket(double val) const
-{
-    if (!active)
-        throw HttpReturnException(500, "Attempt to get bucket from non-numeric value");
-    return std::lower_bound(splits.begin(), splits.end(), val)
-        - splits.begin() + offset;
-}
-
-#if 0
-double
-NumericValues::
-getValue(uint32_t bucket) const
-{
-    if (!active)
-        throw HttpReturnException(500, "Attempt to get value from non-numeric bucket");
-    if (bucket > splits.size())
-        throw HttpReturnException(500, "Attempt to get out of range value from numeric bucket");
-    if (bucket == splits.size()) // shouldn't really need this... TODO to figure it out
-        return INFINITY;
-}
-#endif
-
-uint32_t
-OrdinalValues::
-getBucket(const CellValue & val) const
-{
-    if (!active)
-        throw HttpReturnException(500, "Attempt to get bucket from non-ordinal value");
-    return std::lower_bound(splits.begin(), splits.end(), val)
-        - splits.begin() + offset;
-}
-
-#if 0
-CellValue
-OrdinalValues::
-getValue(uint32_t bucket) const
-{
-}
-#endif
-
-uint32_t
-CategoricalValues::
-getBucket(const CellValue & val) const
-{
-    auto it = std::find(buckets.begin(), buckets.end(), val);
-    if (it != buckets.end())
-        return it - buckets.begin() + offset;
-    throw HttpReturnException(500, "categorical value " + val.toString() + " not found in col");
-}
-
-#if 0
-CellValue
-CategoricalValues::
-getValue(uint32_t bucket) const
-{
-}
-#endif
-
-BucketDescriptions::
-BucketDescriptions()
-{
-}
-
-void
-BucketDescriptions::
-initialize(std::vector<CellValue> values, int numBuckets)
-{
-    std::vector<std::unordered_map<CellValue, size_t> >
-        typeValues(CellValue::NUM_CELL_TYPES);
-    
-    // Segment by type of value
-    for (auto & v: values) {
-        typeValues.at((int)v.cellType())[v] += 1;
-    }
-
-    // Bucketize each type.  Strings can't be bucketized.
-    size_t n = 0;
-    if (!typeValues[CellValue::EMPTY].empty()) {
-        this->hasNulls = true;
-        ++n;  // value zero is for nulls
-    }
-
-    this->numeric.offset = n;
-    this->numeric.active = false;
-    this->numeric.splits.clear();
-
-    // Bucketize numeric values
-    std::vector<CellValue> numerics;
-    for (auto & v: typeValues[CellValue::INTEGER])
-        numerics.push_back(v.first);
-    for (auto & v: typeValues[CellValue::FLOAT])
-        numerics.push_back(v.first);
-    std::sort(numerics.begin(), numerics.end());
-
-    if (!numerics.empty()) {
-
-        std::vector<float> splitPoints;
-        std::vector<std::pair<float, float> > freqPairs;
-        for (auto & n: numerics)
-            freqPairs.emplace_back(n.toDouble(), 1);
-        
-        ML::BucketFreqs freqs(freqPairs.begin(), freqPairs.end());
-        
-        if (numBuckets != -1 && numerics.size() > numBuckets) {
-            ML::bucket_dist_reduced(splitPoints, freqs, numBuckets);
-        } else {
-            ML::bucket_dist_full(splitPoints, freqs);
-        }
-
-        numerics.clear();
-        numerics.insert(numerics.begin(), splitPoints.begin(), splitPoints.end());
-        this->numeric.active = true;
-
-        for (auto & split: numerics) {
-            this->numeric.splits.emplace_back(split.toDouble());
-            ++n;
-        }
-
-        n += 1;
-    }
-
-    // Get string values
-    std::vector<CellValue> stringValues;
-    for (auto type: { CellValue::ASCII_STRING, CellValue::UTF8_STRING } ) {
-        for (auto & v: typeValues[(int)type])
-            stringValues.emplace_back(v.first);
-    }
-    std::sort(stringValues.begin(), stringValues.end(),
-              [&] (const CellValue & v1, const CellValue & v2)
-              {
-                  return v1.toUtf8String() < v2.toUtf8String();
-              });
-
-    this->strings.offset = n;
-    for (auto & v: stringValues) {
-        this->strings.buckets.emplace_back(std::move(v));
-        ++n;
-    }
-
-    // TODO: complete this
-    
-    for (auto type: { CellValue::BLOB, CellValue::TIMESTAMP, CellValue::TIMEINTERVAL } ) {
-        if (!typeValues[(int)type].empty())
-            throw HttpReturnException(500, "Can't bucketize blob, timestamp, interval types");
-    }
-
-    this->blobs.offset = this->timestamps.offset = this->intervals.offset = n;
-    this->blobs.buckets.clear();
-    this->timestamps.splits.clear();
-    this->intervals.splits.clear();
-    this->timestamps.active = this->intervals.active = false;
-
-    ExcAssertEqual(n, this->numBuckets());
-}
-
-#if 0
-    // Finally, perform the bucketed lookup
-    WritableBucketList buckets(totalRows, n);
-
-    for (unsigned i = 0;  i < chunks.size();  ++i) {
-        auto onValue = [&] (size_t, const CellValue & val)
-            {
-                uint32_t bucket;
-                if (val.isNumeric())
-                    bucket = numericValues[val];
-                else bucket = categoricalValues[val];
-
-                ExcAssertLess(bucket, n);
-
-                buckets.write(bucket);
-
-                return true;
-            };
-            
-        chunks[i].columns[it->second].forEach(onValue);
-    }
-#endif
-
-
-CellValue
-BucketDescriptions::
-getSplit(uint32_t bucket) const
-{
-    //return "split";
-    if (bucket < numeric.offset)
-        return CellValue(); //empty
-    else if (bucket < strings.offset)
-        return numeric.splits[bucket - numeric.offset];
-    else if (bucket < blobs.offset)
-        return strings.buckets[bucket - strings.offset];
-
-    throw HttpReturnException(500, "Invalid bucket");
-}
-
-#if 0
-bool
-BucketDescriptions::
-isNumeric() const
-{
-    for (auto & v: values) {
-        if (!v.isNumeric())
-            return false;
-    }
-    return true;
-}
-#endif
-
-bool 
-BucketDescriptions::
-isOnlyNumeric() const
-{
-    return this->strings.buckets.empty(); //todo: complete this
-}
-
-size_t
-BucketDescriptions::
-numBuckets() const
-{
-    return intervals.offset + intervals.numBuckets();
 }
 
 /*****************************************************************************/
