+++ server/server.mk
@@ -26,7 +26,8 @@ LIBMLDB_SOURCES:= \
 	script_output.cc \
 	forwarded_dataset.cc \
 	serial_function.cc \
-    column_scope.cc
+    column_scope.cc \
+    bucket.cc
 
 LIBMLDB_LINK:= \
 	service_peer mldb_builtin_plugins sql_expression git2 hoedown credentials_daemon mldb_builtin command_expression cloud mldb_core
